% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.R
\docType{data}
\name{glasgowComaScale}
\alias{glasgowComaScale}
\title{Glasgow Coma Scale (GCS)}
\format{A data frame}
\usage{
data(glasgowComaScale)
}
\description{
Glasgow Coma Scale (GCS) is a neurological scale which aims to give a reliable and objective way of recording the conscious state of a person for initial as well as subsequent assessment. A patient is assessed against the criteria of the scale, and the resulting points give a patient score between 3 (indicating deep unconsciousness) and either 14 (original scale) or 15 (the more widely used modified or revised scale).
}
\details{
\itemize{
  \item Class Name of GCS
}
}
\references{
https://en.wikipedia.org/wiki/Glasgow_Coma_Scale
}
\keyword{datasets}
